+++ src/components/burger-ingredients/components/burger-ingredients-item/burger-ingredients-item.tsx
@@ -18,7 +18,7 @@ BurgerIngredientsItem.propTypes = {
 export function BurgerIngredientsItem({ part }: IBurgerConstructorItemProps) {
 
 	const dispatch = useDispatch();
-	const state = useSelector((state:RootState) => ({ ...state.main }));
+	const state = useSelector((state:RootState) => ({ ...state }));
 
 	const { ingredientAmountMap } = state;
 	const { price, name, image, _id } = part;
